+++ apps/sim/app/workspace/[workspaceId]/settings/secrets/[credentialId]/components/secret-usage-panel/secret-usage-panel.tsx
@@ -5,8 +5,9 @@ import { ChipLink } from '@sim/emcn'
 import { formatDateTime } from '@sim/utils/formatting'
 import { SettingsActionChip } from '@/components/settings/settings-header'
 import type { SecretUsageEntryPayload, SecretUsageScope } from '@/lib/api/contracts'
+import { DELETED_WORKFLOW_LABEL } from '@/lib/workflows/workflow-labels'
 import { FloatingOverflowText } from '@/app/workspace/[workspaceId]/components'
-import { DELETED_WORKFLOW_LABEL, TriggerBadge } from '@/app/workspace/[workspaceId]/logs/utils'
+import { TriggerBadge } from '@/app/workspace/[workspaceId]/logs/utils'
 import {
   ActivityLog,
   type ActivityLogEntry,
