+++ apps/sim/lib/tokenization/index.ts
@@ -5,14 +5,16 @@ export {
 } from '@/lib/tokenization/calculators'
 export { LLM_BLOCK_TYPES, TOKENIZATION_CONFIG } from '@/lib/tokenization/constants'
 export { createTokenizationError, TokenizationError } from '@/lib/tokenization/errors'
+/**
+ * The exact, `js-tiktoken`-backed counters are deliberately NOT re-exported here.
+ * Re-exporting them would put the tokenizer's 5.4 MB of BPE rank tables back into
+ * every client graph that touches this barrel. Import `@/lib/tokenization/accurate`
+ * directly when an exact count is required.
+ */
 export {
-  batchByTokenLimit,
-  clearEncodingCache,
   estimateInputTokens,
   estimateOutputTokens,
   estimateTokenCount,
-  getAccurateTokenCount,
-  truncateToTokenLimit,
 } from '@/lib/tokenization/estimators'
 export { processStreamingBlockLog, processStreamingBlockLogs } from '@/lib/tokenization/streaming'
 export {
